add Constants and ObjectMapperExtension
